[VLM] Pin pos-embed interpolation linspace to float32
torch.linspace without an explicit dtype follows torch.get_default_dtype().
The vectorized qwen3_vl path and both moss_vl paths omitted it, while the
qwen3_vl loop path pins float32 -- so a polluted global default (e.g. a test
calling torch.set_default_dtype(torch.bfloat16) without restoring) breaks the
documented bit-exact equivalence between the loop and vectorized paths: bf16
quantization of the interpolation coordinates shifts bilinear weights and can
flip floor indices across integer boundaries, gathering different embedding
rows entirely. Pass dtype=torch.float32 everywhere so the interpolation never
depends on process-global state.

+++ python/sglang/srt/models/moss_vl.py
@@ -345,8 +345,12 @@ class MossVLVisionModel(nn.Module):
 
         for _, h, w in zip(grid_ts, grid_hs, grid_ws):
             h_int, w_int = int(h.item()), int(w.item())
-            h_idxs = torch.linspace(0, num_grid_per_side - 1, h_int, device=device)
-            w_idxs = torch.linspace(0, num_grid_per_side - 1, w_int, device=device)
+            h_idxs = torch.linspace(
+                0, num_grid_per_side - 1, h_int, dtype=torch.float32, device=device
+            )
+            w_idxs = torch.linspace(
+                0, num_grid_per_side - 1, w_int, dtype=torch.float32, device=device
+            )
 
             h_idxs_floor = h_idxs.int()
             w_idxs_floor = w_idxs.int()
@@ -454,11 +458,15 @@ class MossVLVisionModel(nn.Module):
             torch.tensor(ws, device=device), return_inverse=True
         )
         h_luts = [
-            torch.linspace(0, num_grid_per_side - 1, int(h), device=device)
+            torch.linspace(
+                0, num_grid_per_side - 1, int(h), dtype=torch.float32, device=device
+            )
             for h in uniq_h.tolist()
         ]
         w_luts = [
-            torch.linspace(0, num_grid_per_side - 1, int(w), device=device)
+            torch.linspace(
+                0, num_grid_per_side - 1, int(w), dtype=torch.float32, device=device
+            )
             for w in uniq_w.tolist()
         ]
         h_lut_off = _exclusive_prefix([len(x) for x in h_luts])
